these are the final changes in code.
We need a Xampp Server - XAMPP 7.4.7 to run this.

This system is an automated tool for internship approval and management system that keeps the record of vital information regarding outhouse internships that students apply for. This project’s main idea is to develop an online centralized application connected to the database which will maintain all the data related to the internship and their status along with necessary required documents.

Safe and secure data: Various tools offered by AIA have enabled the admin members to manage a lot of data related to the internship safely. Important documents and certificates uploaded by students can be viewed only by the admin department and can also be stored as soft copy.